--- a/Project.toml
+++ b/Project.toml
@@ -17,10 +17,3 @@ POMDPs = "0.9"
 POMDPTools = "0.1"
 Parameters = "0.12"
 julia = "1"
-
-[extras]
-POMDPModels = "355abbd5-f08e-5560-ac9e-8b5f2592a0ca"
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-
-[targets]
-test = ["POMDPModels", "Test"]