--- a/Project.toml
+++ b/Project.toml
@@ -8,12 +8,10 @@ version = "0.4.3"
 POMDPLinter = "f3bd98c0-eb40-45e2-9eb1-f2763262d755"
 POMDPTools = "7588e00f-9cae-40de-98dc-e0c70c48cdd7"
 POMDPs = "a93abf59-7444-517b-a68a-c42f96afdd7d"
-Parameters = "d96e819e-fc66-5662-9728-84c9c7592b0a"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 
 [compat]
 POMDPLinter = "0.1"
-POMDPs = "0.9"
 POMDPTools = "0.1"
-Parameters = "0.12"
+POMDPs = "0.9"
 julia = "1"