--- a/src/sarsa_lambda.jl
+++ b/src/sarsa_lambda.jl
@@ -28,7 +28,7 @@ Parameters:
     print information during training
     default: `true`
 """
-@with_kw mutable struct SARSALambdaSolver{E<:ExplorationPolicy} <: Solver
+Base.@kwdef mutable struct SARSALambdaSolver{E<:ExplorationPolicy} <: Solver
    n_episodes::Int64 = 100
    max_episode_length::Int64 = 100
    learning_rate::Float64 = 0.001