For benchmarking, solver configurations are the same as in test files, except `n_episodes = 1_000`.

# Q learning

Before

```julia
BenchmarkTools.Trial: 749 samples with 1 evaluation.
 Range (min … max):  5.293 ms … 10.363 ms  ┊ GC (min … max): 0.00% … 33.24%
 Time  (median):     6.081 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   6.674 ms ±  1.403 ms  ┊ GC (mean ± σ):  9.94% ± 14.36%

      ▂▄▅▆▇█▅▄                                                
  ▂▃▄▇██████████▇▄▃▃▂▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▂▃▃▄▄▆▄▅▅▆▄▄▃▄ ▃
  5.29 ms        Histogram: frequency by time        9.95 ms <

 Memory estimate: 12.36 MiB, allocs estimate: 208948.
```

After

```julia
BenchmarkTools.Trial: 1195 samples with 1 evaluation.
 Range (min … max):  3.196 ms … 9.009 ms  ┊ GC (min … max):  0.00% … 57.34%
 Time  (median):     3.647 ms             ┊ GC (median):     0.00%
 Time  (mean ± σ):   4.181 ms ± 1.529 ms  ┊ GC (mean ± σ):  12.63% ± 17.32%

    ▄▆█▇                                                     
  ▃▆████▇▆▃▂▂▁▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▂▂▃▃▃▃▄▃▃ ▃
  3.2 ms         Histogram: frequency by time       8.62 ms <

 Memory estimate: 8.24 MiB, allocs estimate: 129217.
```

# SARSA

Before

```julia
BenchmarkTools.Trial: 678 samples with 1 evaluation.
 Range (min … max):  5.886 ms … 10.119 ms  ┊ GC (min … max): 0.00% … 22.13%
 Time  (median):     6.815 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   7.369 ms ±  1.094 ms  ┊ GC (mean ± σ):  9.90% ± 11.79%

        ▁▂█▄▂▇▅▂▃▁                                            
  ▂▂▃▄▆▅████████████▅▅▅▄▄▃▃▂▁▂▁▂▁▁▁▁▂▂▁▂▂▃▃▄▆▆▅█▇█▆▆▇▆▆▆▄▃▃▃ ▄
  5.89 ms        Histogram: frequency by time        9.46 ms <

 Memory estimate: 12.55 MiB, allocs estimate: 220650.
```

After

```julia
BenchmarkTools.Trial: 768 samples with 1 evaluation.
 Range (min … max):  5.121 ms … 11.150 ms  ┊ GC (min … max): 0.00% … 40.97%
 Time  (median):     5.900 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   6.509 ms ±  1.558 ms  ┊ GC (mean ± σ):  9.74% ± 14.55%

     ▃▆██▇▅▇▂▁                                                
  ▃▄▇██████████▇▄▃▂▃▁▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▂▁▃▃▃▄▅▅▅▅▄▃▃▄ ▃
  5.12 ms        Histogram: frequency by time        10.5 ms <

 Memory estimate: 10.26 MiB, allocs estimate: 160562.
```

# SARSA-lambda

Before

```julia
BenchmarkTools.Trial: 66 samples with 1 evaluation.
 Range (min … max):  69.143 ms … 82.642 ms  ┊ GC (min … max): 12.24% … 14.93%
 Time  (median):     76.713 ms              ┊ GC (median):    11.49%
 Time  (mean ± σ):   76.025 ms ±  2.978 ms  ┊ GC (mean ± σ):  11.74% ±  1.15%

                              ▅       █  ▅▅   ▂                
  ▅▅▁█▅▅▁▁█▁▁▁▁▁▁▅▁▁▁▅█▅▅▁█▅▅██▁▅▁██▁██▅███▅▅████▅█▅▁▅▁▁▅▁▁▁▅ ▁
  69.1 ms         Histogram: frequency by time        81.4 ms <

 Memory estimate: 160.64 MiB, allocs estimate: 326811.
```

After

```julia
BenchmarkTools.Trial: 86 samples with 1 evaluation.
 Range (min … max):  51.804 ms … 72.022 ms  ┊ GC (min … max): 13.76% … 18.35%
 Time  (median):     58.443 ms              ┊ GC (median):    12.77%
 Time  (mean ± σ):   58.534 ms ±  3.388 ms  ┊ GC (mean ± σ):  14.89% ±  2.56%

                 █▅ ▂   ▂ ▂▂ ▂  ▂▂▂  ▂▂   ▂                    
  ▅▅▁█▁▁▅▁▁▅▁▁▁█▅██▅█▅▅▅█▁████▁▅███▅▁██████▅▅▁▅█▅█▁██▅▅▅▅█▅▁▅ ▁
  51.8 ms         Histogram: frequency by time        64.4 ms <

 Memory estimate: 157.03 MiB, allocs estimate: 263353.
```